extract WebsocketConnector
